iprt: RTPath - lenths are size_t not unsigned.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@14050 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/iprt/path.h
@@ -233,12 +233,12 @@ RTDECL(void) RTPathStripTrailingSlash(char *pszPath);
 
 /**
  * Parses a path.
- * 
- * It figures the length of the directory component, the offset of 
+ *
+ * It figures the length of the directory component, the offset of
  * the file name and the location of the suffix dot.
  *
  * @returns The path length.
- * 
+ *
  * @param   pszPath     Path to find filename in.
  * @param   pcbDir      Where to put the length of the directory component.
  *                      If no directory, this will be 0. Optional.
@@ -248,7 +248,7 @@ RTDECL(void) RTPathStripTrailingSlash(char *pszPath);
  * @param   poffSuff    Where to store the suffix offset (the last dot).
  *                      If empty string or if it's ending with a slash this
  *                      will be set to -1. Optional.
- * @param   pfFlags     Where to set flags returning more information about 
+ * @param   pfFlags     Where to set flags returning more information about
  *                      the path. For the future. Optional.
  */
 RTDECL(size_t) RTPathParse(const char *pszPath, size_t *pcchDir, ssize_t *poffName, ssize_t *poffSuff);
@@ -357,7 +357,7 @@ RTDECL(int) RTPathProgram(char *pszPath, unsigned cchPath);
  * @param   pszPath     Buffer where to store the path.
  * @param   cchPath     Buffer size in bytes.
  */
-RTDECL(int) RTPathUserHome(char *pszPath, unsigned cchPath);
+RTDECL(int) RTPathUserHome(char *pszPath, size_t cchPath);
 
 /**
  * Gets the directory of shared libraries. This is not the same as
@@ -372,7 +372,7 @@ RTDECL(int) RTPathUserHome(char *pszPath, unsigned cchPath);
  * @param   pszPath     Buffer where to store the path.
  * @param   cchPath     Buffer size in bytes.
  */
-RTDECL(int) RTPathSharedLibs(char *pszPath, unsigned cchPath);
+RTDECL(int) RTPathSharedLibs(char *pszPath, size_t cchPath);
 
 /**
  * Gets the directory for architecture-independent application data, for
@@ -386,7 +386,7 @@ RTDECL(int) RTPathSharedLibs(char *pszPath, unsigned cchPath);
  * @param   pszPath     Buffer where to store the path.
  * @param   cchPath     Buffer size in bytes.
  */
-RTDECL(int) RTPathAppPrivateNoArch(char *pszPath, unsigned cchPath);
+RTDECL(int) RTPathAppPrivateNoArch(char *pszPath, size_t cchPath);
 
 /**
  * Gets the directory for architecture-dependent application data, for
@@ -400,7 +400,7 @@ RTDECL(int) RTPathAppPrivateNoArch(char *pszPath, unsigned cchPath);
  * @param   pszPath     Buffer where to store the path.
  * @param   cchPath     Buffer size in bytes.
  */
-RTDECL(int) RTPathAppPrivateArch(char *pszPath, unsigned cchPath);
+RTDECL(int) RTPathAppPrivateArch(char *pszPath, size_t cchPath);
 
 /**
  * Gets the directory for documentation.
